fix(spec-07): align simple_chat protocol
